Implementação inicial de um tratamento de erros para as operações CRUD da entidade Plano

+++ src/main/java/ifsuldeminas/telefonia/controller/comercial/PlanoController.java
@@ -3,9 +3,12 @@ package ifsuldeminas.telefonia.controller.comercial;
 import ifsuldeminas.telefonia.model.entity.comercial.Plano;
 import ifsuldeminas.telefonia.model.repositories.comercial.PlanoRepository;
 import jakarta.validation.Valid;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/planos")
@@ -22,8 +25,12 @@ public class PlanoController {
     }
 
     @GetMapping("/{id}")
-    public Plano get(@PathVariable Long id){
-        return planoRepository.getById(id);
+    public ResponseEntity<Plano> get(@PathVariable Long id){
+        Optional<Plano> opt = planoRepository.findById(id);
+        if(!opt.isPresent()){
+            return new ResponseEntity<Plano>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<Plano>(opt.get(), HttpStatus.OK);
     }
 
     @PostMapping
@@ -32,15 +39,23 @@ public class PlanoController {
     }
 
     @PutMapping("/{id}")
-    public Plano update(@PathVariable Long id, @RequestBody @Valid Plano plano){
-        Plano planoAux = planoRepository.getById(id);
+    public ResponseEntity<Plano> update(@PathVariable Long id, @RequestBody @Valid Plano plano){
+        Optional<Plano> opt = planoRepository.findById(id);
+        if(!opt.isPresent()){
+            return new ResponseEntity<Plano>(HttpStatus.NOT_FOUND);
+        }
+        Plano planoAux = opt.get();
         planoAux.setNome(plano.getNome());
         planoAux.setValorPorMinuto(plano.getValorPorMinuto());
-        return planoRepository.save(planoAux);
+        return new ResponseEntity<Plano>(planoRepository.save(planoAux), HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable Long id){
+    public ResponseEntity<Plano> delete(@PathVariable Long id){
+        if(!planoRepository.existsById(id)){
+            return new ResponseEntity<Plano>(HttpStatus.NOT_FOUND);
+        }
         planoRepository.deleteById(id);
+        return new ResponseEntity<Plano>(HttpStatus.NO_CONTENT);
     }
 }
